% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coa.R
\name{coa}
\alias{coa}
\title{COA Analysis
Params:
  pls_model - estimated seminr PLS model (seminr_model)
  focal_construct - name of focal construct to analyze (character string)
  deviance_bounds - two values of upper and lower deviance threshholds (numeric vector)
  ... - other optional parameters for prediction(), dtree(), or unstable() steps
        e.g., params for unstable()}
\usage{
coa(pls_model, focal_construct, deviance_bounds = c(0.025, 0.975), ...)
}
\description{
COA Analysis
Params:
  pls_model - estimated seminr PLS model (seminr_model)
  focal_construct - name of focal construct to analyze (character string)
  deviance_bounds - two values of upper and lower deviance threshholds (numeric vector)
  ... - other optional parameters for prediction(), dtree(), or unstable() steps
        e.g., params for unstable()
}
