% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rules.R
\name{competes}
\alias{competes}
\title{Reports all competing rules for a node split
Example: (given UTAUT deviant group nodes: 4, 40, 12, 119, 31)
  competes(40, utaut_overfit$dtree)}
\usage{
competes(node_id, dtree)
}
\description{
Reports all competing rules for a node split
Example: (given UTAUT deviant group nodes: 4, 40, 12, 119, 31)
  competes(40, utaut_overfit$dtree)
}
