% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segmentation.R
\name{corr2}
\alias{corr2}
\title{Computes correlation of two matrices of same dimension (from Matlab)
https://stackoverflow.com/questions/27343283/explaining-corr2-function-in-matlab}
\usage{
corr2(A, B)
}
\description{
Computes correlation of two matrices of same dimension (from Matlab)
https://stackoverflow.com/questions/27343283/explaining-corr2-function-in-matlab
}
