% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segmentation.R
\name{model_segments}
\alias{model_segments}
\title{Find segments from unpruned deviance tree}
\usage{
model_segments(coa)
}
\arguments{
\item{coa}{A completed `coa` analysis}
}
\description{
Find segments from unpruned deviance tree
}
