% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segmentation.R
\name{random_split_corrs}
\alias{random_split_corrs}
\title{Simulates random splitting of cases and reports path coefficient correlation}
\usage{
random_split_corrs(original_model, node_cases, node_id)
}
\value{
vector of average correlations between each split coefficients and original model coefficients
}
\description{
Simulates random splitting of cases and reports path coefficient correlation
}
