% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unstable.R
\name{unstable_params}
\alias{unstable_params}
\title{Generates report of unstable parameters for a given pls_model and dtree, or a completed coa analysis
params:
  pls_model - an estimated seminr_model
  dtree - a deviance tree object
  params - parameter, or vector of parameters, to diff}
\usage{
unstable_params(
  pls_model = NULL,
  dtree = NULL,
  analysis = NULL,
  params = "path_coef",
  ...
)
}
\description{
example: (given a seminr_model called utaut_model)
  predictions <- prediction_metrics(utaut_model, "BI")
  deviance_bounds = c(0.025, 0.975)
  dtree <- deviance_tree(pd, deviance_bounds)
  unstable <- unstable_params(
    pls_model, dtree, 
    params = c("path_coef", "outer_weights", "rSquared")
  )
}
